Migrate terminal Effect services
